remove useSpotifyPlayer
