+++ app/src/routes/single-playlist-route.tsx
@@ -1,12 +1,11 @@
-import React, { useCallback, useMemo } from 'react';
+import React, { useMemo } from 'react';
 import { useParams } from 'react-router-dom';
-import { Box, Typography, Grid, Button } from '@mui/material';
+import { Box, Typography, Grid } from '@mui/material';
 import { PlaylistList } from '../app/playlist-list';
 import { useUserPlaylists } from '../hooks/use-user-playlists';
 import { useTracksForPlaylist } from '../hooks/use-tracks-for-playlist';
 import { BaseRoute } from './base-route';
 import { TracksTable } from '../app/tracks-table';
-import { useSpotifyPlayer } from '../hooks/use-spotify-player';
 
 export function SinglePlaylistRoute() {
   const { playlists } = useUserPlaylists();
@@ -25,16 +24,6 @@ export function SinglePlaylistRoute() {
 
   const playlistUri = currentPlaylist?.uri;
 
-  const { play } = useSpotifyPlayer();
-
-  const handleClickPlay = useCallback(() => {
-    if (!playlistUri) {
-      return;
-    }
-    play(null, playlistUri);
-    console.log('play', null, playlistUri);
-  }, [play, playlistUri]);
-
   return (
     <BaseRoute>
       <Box sx={{ pt: 8, height: '100vh' }}>
@@ -46,7 +35,6 @@ export function SinglePlaylistRoute() {
           <Grid item xs={9}>
             <Box>
               <Typography variant="h4">{currentPlaylist?.name}</Typography>
-              <Button onClick={handleClickPlay}>Play Playlist</Button>
             </Box>
             <TracksTable tracks={tracks} contextUri={playlistUri} />
           </Grid>
